changed smallint to int

+++ sql/setup-schema.sql
@@ -1,4 +1,4 @@
 --create user table
-create table Users(id smallint primary key, username varchar(25), avatar varchar(15), follows varchar(25));
-create table Tweets(id smallint primary key, user_id smallint, message varchar(140), hit_time timestamp, foreign key(user_id) references Users(id));
-create table Followers(id smallint primary key, user_id1 smallint, user_id2 smallint, foreign key(user_id1) references Users(id), foreign key(user_id2) references Users(id));
+create table Users(id int primary key, username varchar(25), avatar varchar(15), follows varchar(25));
+create table Tweets(id int primary key, user_id int, message varchar(140), hit_time timestamp, foreign key(user_id) references Users(id));
+create table Followers(id int primary key, user_id1 int, user_id2 int, foreign key(user_id1) references Users(id), foreign key(user_id2) references Users(id));
